+++ CHANGELOG.md
@@ -13,5 +13,6 @@
 - Fixed the Clean type doing the opposite of what it is supposed to do.
 - Fixed the Leafblower type not dropping items as expected, and fixed its complete failure when used in a size that
   would make it executed asynchronous.
+- Fixed an error that would occur when using a brush with sphere or cylinder shape, with a size of 0.
 
 ### For PocketMine-MP API: 4.0.0